requirements.txt
textual>=0.59.0
ytmusicapi>=1.6.0
python-mpv>=1.0.6
yt-dlp>=2024.12.13
sounddevice>=0.5.0
numpy>=2.1.3
